requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
requests
os
matplotlib.pyplot
pickle
seaborn 
warnings
